reorganize solidago modules
