+++ solidago/src/solidago/comparisons_to_scores/base.py
