+++ solidago/src/solidago/comparisons_to_scores/indiv_matrix_inversion.py
