--- solidago/src/solidago/mehestan/__init__.py
@@ -1,1 +1,0 @@
-from .primitives import QrMed, QrUnc, QrDev
