+++ solidago/src/solidago/resilient_primitives/__init__.py
@@ -1,0 +1,1 @@
+from ._primitives import QrMed, QrDev, QrUnc, BrMean
