+++ solidago/src/solidago/solvers/optimize.py
