+++ solidago/tests/test_resilient_primitives.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from solidago.mehestan.primitives import QrDev, QrMed, QrUnc
+from solidago.resilient_primitives import QrDev, QrMed, QrUnc
 
 
 class QrMedTest(TestCase):
@@ -16,8 +16,8 @@ class QrMedTest(TestCase):
                 x=np.array([-10.0, 1.0, 10.0]),
                 delta=np.array([1e-3, 1e-3, 1e-3]),
             ),
-            0.5, # influence of the median contributor is bounded by weight / W
-            places=3
+            0.5,  # influence of the median contributor is bounded by weight / W
+            places=3,
         )
 
     def test_qrmed_with_W_equals_zero(self):
@@ -32,31 +32,31 @@ class QrMedTest(TestCase):
                 delta=np.array([1e-3, 1e-3, 1e-3]),
             ),
             1.0,
-            places=3
+            places=3,
         )
 
     def test_qrmed_with_high_uncertainty(self):
         """Each score with a high uncertainty will have less effect on QrMed."""
         self.assertAlmostEqual(
             QrMed(
-                W=1.,
-                w=1.,
-                x=np.array([-10., 1., 10.]),
-                delta=1000000.,
+                W=1.0,
+                w=1.0,
+                x=np.array([-10.0, 1.0, 10.0]),
+                delta=1000000.0,
             ),
-            0.,
-            places=5
+            0.0,
+            places=5,
         )
 
 
 class QrDevTest(TestCase):
     def test_qrdev_default(self):
-        default_deviation = 3.
+        default_deviation = 3.0
         self.assertEqual(
             QrDev(
-                W=2.,
+                W=2.0,
                 default_dev=default_deviation,
-                w=1.,
+                w=1.0,
                 x=np.array([]),
                 delta=np.array([]),
             ),
@@ -64,29 +64,29 @@ class QrDevTest(TestCase):
         )
 
     def test_qrdev_with_high_weight(self):
-        x = np.array([-6., 0., 15.])
+        x = np.array([-6.0, 0.0, 15.0])
         self.assertAlmostEqual(
             QrDev(
-                W=2.,
-                default_dev=1.,
-                w=100000.,
+                W=2.0,
+                default_dev=1.0,
+                w=100000.0,
                 x=x,
-                delta=0.,
+                delta=0.0,
             ),
             float(np.median(np.abs(x))),  # = 6.
-            places=5
+            places=5,
         )
 
     def test_qrdev_with_W_equals_0(self):
         """Setting W to 0 is like removing the bias towards default_dev"""
-        x = np.array([-6., 0., 15.])
+        x = np.array([-6.0, 0.0, 15.0])
         self.assertAlmostEqual(
             QrDev(
-                W=0.,
-                default_dev=1.,
-                w=1.,
+                W=0.0,
+                default_dev=1.0,
+                w=1.0,
                 x=x,
-                delta=0.,
+                delta=0.0,
             ),
             float(np.median(np.abs(x))),  # = 6.
             places=5,
@@ -99,17 +99,17 @@ class QrDevTest(TestCase):
         thus not change much from its default deviation.
         """
         default_deviation = 3
-        x = np.array([-6., 0., 15.])
+        x = np.array([-6.0, 0.0, 15.0])
         self.assertAlmostEqual(
             QrDev(
-                W=2.,
+                W=2.0,
                 default_dev=default_deviation,
-                w=1.,
+                w=1.0,
                 x=x,
-                delta=100000000.,
+                delta=100000000.0,
             ),
             default_deviation,
-            places=5
+            places=5,
         )
